First version of new AC 156 Suggested Tip Rate dialog.
Since the code is so similar to AC154 for labor rate, we are just putting the AC156 dialog logic into AC154 dialog logic source code file.

These are the UI changes that present data to the users.

Still to do are the changes to functions that interface with the memory resident database (ParaAllWrite(), etc.) and the actual retrieving and storing of data.

Also needed are specific language DLL changes to modify the displayed dialog changing from English to the target language.