set AC162, Flexible Drive Through, defaults to be for six counter terminals (#1 through #6) with no Drive Thru terminals.
this parameter data is only used for a Store/Recall System for Quick Serivce Restaurant however it is easy to forget that this parameter exists and must be set when Store/Recall is being used for something other than a Quick Service Restaurant.

also updated some comments to make things easier to find using search.