in function TtlReqBackEachPlu() initialize the buffers to zero rather than using memset() and use RflGetPtdFlagByType(FLEX_PLU_ADR) to get the Period To Date flag for PLU totals rather than a call to CliParaRead() as part of just making the source a bit simpler.
Also if PluTotalDelFile() fails then issue an ASSRTLOG log that explains the other ASSRTLOG logs generated by the called function tree when doing a delete.

Moved the additional initialization of RequestBuff directly above the loop which is then going to use it to iterate over the PLU totals database when making a copy for synchronization between Master and Backup Master terminals.