testing printing receipt with a direct to COMM port receipt printer using an NCR 7198 printer resulted in the printer not following the align center and the align right commands when multiple text justification commands were used on the same print line.
reading the manual, it appears that though having mutilple text justification commands in the same print line worked with OPOS, it does not seem to be allowed by the printer firmware itself.

we have created a new function the preprocesses the print line to perform text adjustment removing the align center and the align right text justification commands and creating a line of text using spaces to justify the text.

WARNING: it looks like the direct to COMM port functionality needs to change the line lenght used by the NCR 7198 receipt printer. it appears that the line length is currently 44 characters and it should be at least 48 characters.