if the historical report functionality fails to open the html output file, handle the error appropriately with ASSRTLOG logs to indicate the failure and to not just crash GenPOS.
A customer had created a historical reports folder path that had a space after the end. The result was that the pathname was reports followed by a space rather than reports with no space and the target folder was named reports with no space. this resulted in a file open error that caused GenPOS to crash.

We now log the error and also abort the historical report generation should there be a file open error.