For the historicalreportsfolder tag when processing the PARAMINI file we will trim leading and trailing spaces. We are doing this because people will sometimes have a leading or trailing space in order to improve readability.
This has come up when investigating a problem with GenPOS crashing when using the historical reports functionality which was due to not handling a file open error correctly. The user had inserted a trailing space and the resulting pathname with the trailing space was incorrect.

Having a pathname with a trailing space doesn't seem to have much utility.