improve the historical reports functionality to allow multiple different report formats: html, csv, etc.
With these changes we have implemented AC21 report engine changes and have made a minor change to the EOD and AC23 reports so that the HTML reports they currently generate will work with this new functionality.

This is still a work in progress and needs further development however it currently compiles and does generate reports.