update the historical reports generation functions to allow for various file types.  we are making changes to support multiple output file types including CSV, XML, HTML, and the Hubworks file type used with Hubworks.com.
we are concentrating on using AC21 and AC23 reports with these changes for now with the goal of extending the functionality for all of the reports.

in addition to historical reports folder which is intended to be a folder connect to a cloud backup interface, we are also making provisions for temporary reports to be generated in a special folder to allow viewing with a web browser.