regularize the interface for the ItemGenerateAc21Report() function and add the ItemGenerateAc30Report() function for coupons.
we still have work to do with the coupon report functionality.

we still have work to refactor the functionality of the various reports and to support all of the report types: html, csv, json, txt, xml.

we need to try to consolidate the output function so that the same output function used by AC21 and AC23 reporting functionality can be used with other reports.

we need to update the data query functionality to allow the same reports through it as well. right now only AC21 and AC23 are available.