refactor the historical reports generating code for AC 21, AC 23, and AC 30 so that the End Of Day report as well as the individual reports are created properly.
we still have an issue with the AC 23 financial report in that the EOD report repeats terminal 1 as terminal 0 and the AC 23 financial report created has terminal 0 but not terminal 1. it should be terminal 1.

still need to do testing for the printed reports to ensure the printed reports are still generated correctly.

this does raise an issue about the Supervisor Mode and reports. currently historical reports are generated when the AC report entry is done in Supervisor Mode so that new historical reports are always generated at the same time a printed report is generated.

do we need to decouple these two functions?