now starting to integrate DataQuery with the standard Supervisor Mode AC reports. we have a new report folder type, RptElementOutputFolder,  which allows us to specify either a standard target folder such as the web server folder or the queries folder or to use the PARAMINI designated historical reports folder.
we've added a new standard folder, STD_FOLDER_QUERYFOLDER, which is created at GenPOS startup if needed.

we need to differentiate between what are meant to be persistent, historical data files, and what are temporary files that are used for a specific data query or for a report which is then printed.

we've made a couple of tweaks to the RptDescription functionality to better integrate DataQuery with the Supervisor Mode reports. we've also made tweaks to the ItemGenerateAc functions for AC21, 23, and 30 so that the same functions can be used with the AC99 End Of Day report functionality.

and we've updated and improved comments in various places.