For some reason the P08 dialog and supporting source code was modified to remove the list of Supervisor IDs.
The dialog changes were not made in the Chinese or other language resource files, only the English resource file was modified.

In order to display the available Supervisor IDs, I have ported the P08 dialog content which displays the list of 20 Supervisor IDs and modified the P08 dialog initialization routine to allow the user to see the list of available Supervisor IDs that can be used with AC20 Operator Maintenance to create an Operator profile that can be used in Supervisor Mode.

The ability to change the Supervisor IDs is not available but I wanted to get the ability to at least see the list of Supervisor IDs available to the user.