while reading code found an error with incorrectly updating transaction data of a paid guest check that has been recalled.
Defect in function TrnSetConsNoToPostR().

The consecutive number that was being updated used an incorrect calculation for the offset to where the consecutive number was stored.

seen only with a store recall system when recalling a paid out check.

The result was an electronic journal entry with incorrect consecutive number.

defect introduced back in 2011 or so with a change to the struct TRANGCFQUAL which moved the consecutive number offset from the first member of the struct.

also changes to use type TrnFileSize rather than USHORT for the file size data in the transaction file.