we are starting to use const where appropriate and to replacing some of the defines such as FARCONST that are no longer used. We are also using the typedefs such as PifFileHandle in place of intrinsic types such as short.
We are replacing several special functions with the common function they use. Replacing MldCalStoSize() with TrnCalStoSize(), MldExpandFile() with TrnExpandFile().

we eliminated the unused and ifdefed out TrnReadFile_MemoryForce() code that was intended to allow for keeping the transaction data in memory. A defect caused it to be ifdefed out along with the other memory resident transaction file source code to return to the original code of using a disk file. It's time to go ahead and remove detritus that has been hanging about since 2015 or so.