modifying portions of the Client, Server, and Isp subsystems that handle the network messaging between terminals as well as between terminal utilities such as PEP or PCSample.
The main thrust of these changes is to prepare for a change from a USHORT to a ULONG for file transfer sizes.

See https://github.com/opengenpos/genpos/issues/13

However as part of testing the current state of changes, we have found some networking issues with error recovery when sending guest checks from the master terminal to a satellite. Most of that testing has been with drive thru on a two terminal system with order/payment on terminal 1 and delivery on terminal 2.

We are also modifying the function PifGetDateTimeTimer() to accept a PIFTIMER argument rather than a DATE_TIME argument to eliminate the need for conversion with timers.

The purpose of these changes is to make further changes to handle the transmission of larger data files, larger than 0xffff or the maximum for an unsigned short, which is currently a problem in GenPOS.

This problem with file size is most commonly experienced with Guest Checks whose size exceeds the maximum due to the max number of items provisioned.

The size of a Guest Check varies depending on System Type (Guest Check Buffering/Unbuffering, Store Recall, etc.) and the max number of items set in the P2 Flexible Memory settings.

In addition we are starting to use const where appropriate and to replacing some of the defines such as FARCONST that are no longer used. We are also using the typedefs such as PifFileHandle in place of intrinsic types such as short.