corrected some issues with using the SQLite ADODB control and SQLite specific SQL differences.
database in SQLite is a file containing tables and is not a standard SQL object so SQL statements such as CREATE DATABASE don't work. we did have to modify the logic in CnPluTotalDb::CreateDbFile() to open the file for reading to check if it exists and if the open fails then open with create and write modes to create it.

a work around was need to determine if a table exists which a Stack Overflow post provided the SQL for doing so.

in CnPluTotalDb::DropTable() check that the table exists before trying to drop it to prevent misleading ASSRTLOG logs.

a better understanding of the Dollar Tree scer and the work done to back up the PLU database on startup led to a few changes and the removal of a misleading NHPOS_ASSERT().

added some additional logs that are enabled by a defined Preprocessor constant CNPLUTOTALDB_LOG which is currently set to zero to turn them off.

updated a few comments.