remove the define ParaFSC to (Para.ParaFSC) and find the places in the source where the ParaFSC or keyboard mapping data located in the UNINIRAM memory resident database is referenced as ParaFSC and change the reference to the full name, Para.ParaFSC.
The memory resident provisioning data is stored in a global UNINIRAM variable Para.

A few places where functions return lead thru mnemonic addresses so change the return type to USLDTERR to indicate the function returns a lead thru mnemonic.

in function UifRegPauseString() we found a warning about the comma operator which had to be changed to the logical AND operator.

from
  for( i = 0; i < usSize, auchWork[(*puchPtr) - 1][i] !=0; i++) {

to
   for( i = 0; i < usSize && auchWork[(*puchPtr) - 1][i] !=0; i++) {