replace TrnTaxSystem() and PrtTaxSystem() internals with a call to ItemCommonTaxSystem().
This is a another step in eliminating duplication through the source code base.

we are keeping the old defined constants for now however they use the same values as the corresponding ITM_TAX_ defined constants.

also renumbered the ITM_ defined constants in itmlocal.h so they are all negative with unique values to be used with the new typedef of SLDTITM. this renumbering making the ITM_ constants negative numbers makes for two separate ranges, one for lead thru mnemonic addresses and one for the ITM_ defined constants.