refactor and correct a few issues in the adjective modify code.
we found a large block of source that was never being executed since it was after a return statement. this appears to be due to a copy paste that wasn't cleaned up.

make sure all of the various error return values are correct and appropriate. using the new function return types SUIFRSLT and SLDTITM to clarify the values that may be returned by functions.

the adjective modify OEP window was not responding properly to the NONE/DONE button and we made sure that the Cancel key can also be used to close the window.