Several defines and extern declarations used in both MLD as well as Print subsystems moved to prt.h in BusinessLogic/Include.
there is some commonality in function between MLD and Print subsystems and some functions are duplicated between the two. see functions PrtThrmSupMakeETKFl() and MldMaintMakeETKFl().

modified struct RPTEMPLOYEE to use ETK_FIELD rather than a list of the ETK_FIELD fields/members makeing it more similar to struct MAINTETKFL. this change riplled into subsystems Mld/DISPSUP and Report

over the years print functions have changed as has the C compiler standard. one large change was going from a jounal printer to an electronic journal.

one code pattern is to have a main function for a print action that then uses a printer type specific function to perform the action such as the following
 - PrtCoupon
 - PrtCoupon_EJ
 - PrtCoupon_RJ
 -  PrtCoupon_SP
 -  PrtCoupon_TH
 -  PrtCoupon_VL

changed to make the printer type specific functions static with file scope visibility and removed declarations from include files.

a large number of small changes through the Print subsystem

first goal of these changes was to eliminate compiler warnings some of which pointed to defects to be corrected such as buffer overrun warnings. a lot of this was initializing variables when defined and removing declarations in include files for no longer existing functions. several buffer overrun possibilities eliminated.

second goal was to localize variables.

third goal was to initialize variable when defined and eliminate memset() calls.

fourth goal was to eliminate declarations for lower level functions that were only used in a single place and consolidate related functions in the same source file.

fifth goal was to improve readability of the source, add comments, use const for function arguments where appropriate,, add NHPOS_ASSERT() where appropriate, add a few compile time checks on variable sizes.

in several functions that created a dummy TRANINFORMATION that was then populated with a couple of data items for a following function, the compiler issued a warning about the amount of stack space so it was made a static with function scope. This needs to be updated so that we don't have several of these data chunks floating about taking up room.