testing with a receipt printer we discovered receipts were not printed nor put into Electronic Journal. we discovered a defect inserted by an update to file prctran_.c in commit e0b29e78 made Aug-01-2025 introduced a defect in which ITEMPRINT classes were not being processed.
new function PrtSetPrintCompulMask() to allow changes to fsPrtCompul and fsPrtNoPrtMask both of which are global variables in the printing functionality for controlling which print station will be used when printing. reading the code, it appears that these variables are reset anyway and so any changes made using this funcition will be overwritten however the original code was accessing these globals directly and we wanted to require the use of a helper function instead.

remove the two globals with the same names defined in file uireg.c which are seen as declarations rather than definitions since no value is assigned to them.

modified places outside of the Print subsystem and library accessing these globals directly to use the new function instead.