eliminate unused funcion PluTotalNextRec() and the underlying NextRec() methods on which it depends.
correct problem with End of Day which required additional changes for transition to embedded SQLite.

new function GetValuesFromRec() added for pulling values from the current database record and populating the SAFEARRAY with the values. This function is used with  CnPluTotalDb::GetRec() and CnPluTotalDb::GetRow().

NHPOS_NONASSERT_TEXT() logs added to be able to track progress through ASSRTLOG messages. these should be removed before moving to production.