a few tweaks to the main PEP window. using a light gray color to the background instead of the ghastly white.
repositioned the text area used for displaying the basic sizes of the currently loaded PEP file to provide a scrunch more room for the file pathnames.

thinking about changing the font size for the supported Western languages, especially for English from 8 to 10 to make it a bit more readable. however that may make the dialogs a bit wonky.