change the font of the user interface to use Segoe UI at 9 pt. This has required changing to owner drawn menu items.
in PEP.H there are a couple of defines that allow the font to be changed and the source recompiled with the new font and font size.

changes to the menus to make them less cumbersome.

changes to dialogs to update the look and make them more usable and more pleasing to the eye.

in P02 removed the memory usage graphic left over from the old NHPOS 1.4 days. it's not useful anymore and hasn't been since the change to Rel 2.0 and Windows XP.