remove the P54 and P62 dialogs from PEP menu since that functionality is no longer needed.
P54 is EPT and Hotel Charge Post setup. EPT setup is now done on a tender key basis in the P98 Tender Key Setup dialog.

P62 is the tender key setup for assigning a total key to a tender key. this is also now done in the P98 Tender Key Seup dialog.

cleaned up the P10 and P15 dialogs as well as the P50 and P51 dialogs.

Added additional heading and prompt about the indicator for number of leading digits in P10 dialog.

P15 dialog data entry now responds to the MDC 82 bit A setting for number of digits entered as well as prompt about data limits.

some cleanup of source in paraall.c with removal of unneeded FAR keyword and added a test for NULL pointer to eliminate compiler warnings.

new strings added to resource files.